Abandon git-historian for now.
The previous algorithm for gathering history (see
https://crates.io/crates/git-historian) was incorrect.
It walked all previous commits, tracking files by their names as it
traveled backwards through history. Unfortunately, this goes poorly
as soon as merges get involved.
If we see a merge, how do we know how much (if at all) each branch
affected the contents of a file? Without tracking content on a
deeper level, there's no way to know.

To rememdy this, copyrighter now just calls

    git log --follow -M -C

on each file. This is much slower than a single traversal of the
history graph, but one can assume Git does a better job of gathering
a file's history.